Fix issues with multiple instances...
...by removing the control parameters from the Unity AudioEffect registry.

This means we need to make a new API method to change parameters on a
per-synth-instance basis, and add a CS class to manage that.

In the end, I'm hoping it comes out cleaner than Helm's instance
handling was.

Another thing that needs to be fixed is if two OvenMit AudioEffects both
receive audio from the same synth instance, the synth instance is
`proceed`ing twice.