Plugin: Add pan ability to synth instances.
Allow Unity to specify that a particular synth instance should pan its
output. This supports mono and stereo panning modes (mono just fades to
the left and right, and stereo uses a 2x linear fade, so that the middle
amplitude is the same in either modes).

More than 2 channels is not yet supported, and not high priority right
now.